fix(timeline): ripple-delete undo restores track; command clips get real ids
TDD cycle 1. The failing test exposed two coupled defects:
- AddClipCommand never propagated assigned ids (all command-added clips had
  id 0), so rippleDeleteRaw's removeIf(id==deleted) removed EVERY clip on
  the track at redo time. insertClipRaw now assigns ids when id==0 and
  returns the id used; AddClipCommand carries it across redo cycles.
- RippleDeleteCommand::undo restored an empty snapshot (captureSnapshot was
  never called by anyone) -> undo wiped the whole track. Snapshot is now
  taken in redo() before mutation; undo restores it verbatim.
Removed the dead captureSnapshot() hook.

Test: rippleDeleteUndoRestoresTrack (RED->GREEN verified), suite 4/4.

+++ src/model/timeline_datamodel.h
@@ -54,8 +54,8 @@ public:
     QUndoStack* undoStack() { return &m_undo; }
 
     // Direct mutation — ONLY for command internals and tests. Not an edit API.
-    // Returns index of inserted clip on the track.
-    int insertClipRaw(int track, const Clip& c);
+    // Inserts c; if c.id == 0 a fresh id is assigned. Returns the id actually used.
+    qint64 insertClipRaw(int track, const Clip& c);
     bool removeClipRaw(int track, qint64 clipId);
     // Replace the entire contents of a track with the given clips
     // (sorted-by-start snapshot). Command undo support.
